+++ frontend/components/navigation/hire-me.php
@@ -1,1 +1,3 @@
-<?= c('navigation/link', [ 'label' => 'Hire me!', 'href' => 'https://jofaval.com/mlai/contact', 'class' => 'contact-link' ]); ?>
+<?php if (!isset($label)) $label = 'Hire me!' ?>
+
+<?= c('navigation/link', [ 'label' => $label, 'href' => 'https://jofaval.com/mlai/contact', 'class' => 'contact-link' ]); ?>